+++ flight/drone/drone_controller.py
@@ -9,15 +9,12 @@ import traceback
 from drone import Drone
 import exceptions
 
-from flight import constants as c
-from flight.tasks.exit_task import ExitTask
-from flight.tasks.hover_task import HoverTask
-from flight.tasks.land_task import LandTask
-from flight.tasks import HoverTask, TakeoffTask, ExitTask, LinearMovementTask, LandTask
-from flight.tasks.takeoff_task import TakeoffTask
+import flight.constants as c
+import flight.tasks as tasks
 from flight.utils.priority_queue import PriorityQueue
 from flight.utils.timer import Timer
 from flight import flightconfig as f
+from tools.data_distributor.data_splitter import DataSplitter
 
 
 SAFETY_CHECKS_TAG = "Safety Checks"
@@ -133,7 +130,7 @@ class DroneController(object):
         priority : Priorities.{LOW, MEDIUM, HIGH}, optional
             The importance of this task.
         """
-        new_task = HoverTask(self._drone, altitude, duration)
+        new_task = tasks.Hover(self._drone, altitude, duration)
         self._task_queue.push(priority, new_task)
 
     def add_takeoff_task(self, altitude, priority=c.Priorities.HIGH):
@@ -150,7 +147,7 @@ class DroneController(object):
         -----
         Internally, the priority of this task is always set to HIGH.
         """
-        new_task = TakeoffTask(self._drone, altitude)
+        new_task = tasks.Takeoff(self._drone, altitude)
         self._task_queue.push(priority, new_task)
 
     def add_linear_movement_task(
@@ -166,7 +163,7 @@ class DroneController(object):
         priority : Priorities.{LOW, MEDIUM, HIGH}, optional
             The importance of this task.
         """
-        new_task = LinearMovementTask(self._drone, direction, duration)
+        new_task = tasks.LinearMovement(self._drone, direction, duration)
         self._task_queue.push(priority, new_task)
 
 
@@ -178,7 +175,7 @@ class DroneController(object):
         priority : Priorities.{LOW, MEDIUM, HIGH}, optional
             The importance of this task.
         """
-        new_task = LandTask(self._drone)
+        new_task = tasks.Land(self._drone)
         self._task_queue.push(priority, new_task)
 
     def add_exit_task(self, priority=c.Priorities.HIGH):
@@ -188,7 +185,7 @@ class DroneController(object):
         -----
         Always has high priority
         """
-        new_task = ExitTask(self._drone)
+        new_task = tasks.Exit(self._drone)
         self._task_queue.push(priority, new_task)
 
     def _update(self):
@@ -208,7 +205,7 @@ class DroneController(object):
                 # We are done with the task
                 self._logger.info('Finished {}...'.format(
                     type(self._current_task).__name__))
-                if isinstance(self._current_task, ExitTask):
+                if isinstance(self._current_task, tasks.Exit):
                     return False
                 self._task_queue.pop()
