+++ run_AI.py
@@ -1,0 +1,3 @@
+from flight.AIs import user_command
+
+user_command.main()
